Completely restructured factions to use faction IDs, not array indexes.
Array indexes were chosen back when the faction stack was unchanging,
but the existence of dynamic factions makes this system dangerous now,
and likely connected to the errors I've seen with dynamic factions.
This new system takes a page from pilots, giving factions a unique
unsigned long ID which is used to identify them. This will prevent
things like one dynamic faction being treated the same as a previously
deleted dynamic faction.

Miraculously, this change alone seems to have fixed the bug with
seeming ghost references to deleted dynamic factions I witnessed
before, which is good news! I thought it probably had something to
do with that flimsy system (hence why I went through all this effort
to make the switch), but I didn't know it was the *direct* cause.
Looks like I can breathe a sigh of relief (at least, assuming there
aren't more memory access errors that I missed).

+++ src/info.c
@@ -68,7 +68,7 @@ static unsigned int info_wid = 0;
 static unsigned int *info_windows = NULL;
 
 static CstSlotWidget info_eq_weaps;
-static int *info_factions;
+static unsigned long *info_factions;
 
 static int selectedLog = 0;
 static int nlogs=0;
