+++ src/pilot.h
@@ -221,7 +221,7 @@ typedef struct Pilot_ {
    char* name;       /**< pilot's name (if unique) */
 
    /* Fleet/faction management. */
-   int faction;      /**< Pilot's faction. */
+   unsigned long faction; /**< Pilot's faction. */
    int presence;     /**< Presence being used by the pilot. */
 
    /* Object characteristics */
@@ -445,14 +445,15 @@ PilotOutfitSlot* pilot_getDockSlot( Pilot* p );
 /*
  * creation
  */
-unsigned long pilot_create(const Ship* ship, const char* name, int faction,
-      const char *ai, const double dir, const Vector2d* pos,
-      const Vector2d* vel, const PilotFlags flags, unsigned long dockpilot,
-      int dockslot);
-Pilot* pilot_createEmpty( const Ship* ship, const char* name,
-      int faction, const char *ai, PilotFlags flags );
+unsigned long pilot_create(const Ship* ship, const char* name,
+      unsigned long faction, const char *ai, const double dir,
+      const Vector2d* pos, const Vector2d* vel, const PilotFlags flags,
+      unsigned long dockpilot, int dockslot);
+Pilot* pilot_createEmpty(const Ship* ship, const char* name,
+      unsigned long faction, const char *ai, PilotFlags flags);
 Pilot* pilot_replacePlayer( Pilot* after );
-void pilot_choosePoint( Vector2d *vp, Planet **planet, JumpPoint **jump, int lf, int ignore_rules, int guerilla );
+void pilot_choosePoint(Vector2d *vp, Planet **planet, JumpPoint **jump,
+      unsigned long lf, int ignore_rules, int guerilla);
 void pilot_delete( Pilot *p );
